Introduce switch to enable doing a token exchange for signing
Introduce TARGET_HF_REPO_DO_TOKEN_EXCHANGE switch to enable a token
exchange of the GitHub token against and IBM sigstore signing token so
that the model signature appears to be from an IBM account rather than
the github identity that was used to run the build.

Note that not everyone can sign with the IBM sigstore signing token since
it requires per-user setup by 'me'.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>